pyproject.toml
[tool.poetry]
name = "microservices-locations"
version = "0.1.0"
description = ""
authors = ["Daniel Suniaga <danieldsuniaga@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
asgiref = "3.5.2"
astroid = "2.12.10"
certifi = "2022.9.24"
charset-normalizer = "2.1.1"
dill = "0.3.5.1"
Django = "4.1.2"
djangorestframework = "3.13.1"
idna = "3.4"
isort = "5.10.1"
lazy-object-proxy = "1.7.1"
mccabe = "0.7.0"
platformdirs = "2.5.2"
pylint = "2.15.3"
PyMySQL = "1.1.0"
pytz = "2022.4"
requests = "2.28.1"
setuptools = "65.3.0"
six = "1.16.0"
sqlparse = "0.4.3"
tomli = "2.0.1"
tomlkit = "0.11.5"
typing_extensions = "4.3.0"
urllib3 = "1.26.12"
websocket-client = "0.56.0"
wheel = "0.37.1"
wrapt = "1.14.1"
python-decouple = "3.5"
cryptography = "2.7"
celery = "5.2.7"
redis = "4.3.4"
structlog = "22.1.0"
coverage = "6.4.1"

[tool.poetry.dev-dependencies]
